Added three sub-pages to christianity/marriage.xml
I don't think there's such thing as an impartial source for this sort of thing; the people who know the Bible well enough to be able to make judgements based on it are largely people who, in studying it so much, have formed their own opinions about it. :-( I guess I'll just have to cite a lot of them.